port from utf8

+++ cell.h
@@ -26,8 +26,8 @@
 class	CSerial
 {
 public:
-	int	tindex;						// 演算子のCCell位置番号
-	std::vector<int> index;				// 演算対象のCCell位置番号のリスト
+	size_t	tindex;						// 演算子のCCell位置番号
+	std::vector<size_t> index;				// 演算対象のCCell位置番号のリスト
 									// 演算子がF_TAG_FUNCPARAM/F_TAG_SYSFUNCPARAMの場合は、index[0]が関数を示します
 public:
 	CSerial(int t) { tindex = t; }
@@ -42,8 +42,8 @@ class	CCell
 {
 public:
 	yaya::string_t	name;			// この項の"名前"（thisがローカル変数/関数の時に使用します）
-	int		index;					// 位置番号（thisが変数/ローカル変数/関数の際に使用します）
-	int		depth;					// {}入れ子の深さ（thisがローカル変数の時に使用します）
+	size_t		index;					// 位置番号（thisが変数/ローカル変数/関数の際に使用します）
+	size_t		depth;					// {}入れ子の深さ（thisがローカル変数の時に使用します）
 
 private:
 	int     m_type;									// m_valueにタイプしか格納しない場合に使用
